+++ src/lib/isMACAddress.js
@@ -5,7 +5,7 @@ const macAddressNoColons = /^([0-9a-fA-F]){12}$/;
 
 export default function isMACAddress(str, options) {
   assertString(str);
-  if (options && options.noColons) {
+  if (options && options.no_colons) {
     return macAddressNoColons.test(str);
   }
   return macAddress.test(str);